PHASE 22 final corrective: bounded assistant agent and test hygiene.
Enforce per-turn tool budget, transactional tool sessions with deferred embeddings,
bounded model tool output, and evidence-separated UI context; show affected_objects
in Flutter; fix test isolation; green full pytest/ruff/flutter suite.

+++ backend/app/services/bounded_chunks.py
@@ -34,7 +34,7 @@ def select_bounded_chunk_indices(total_chunks: int, max_chunks: int) -> list[int
     indices: list[int] = []
     seen: set[int] = set()
     for slot in range(max_chunks):
-        index = int(round(slot * (total_chunks - 1) / (max_chunks - 1)))
+        index = round(slot * (total_chunks - 1) / (max_chunks - 1))
         if index not in seen:
             seen.add(index)
             indices.append(index)
